Refactor how branches are managed with PCRProtectionProfile.
This removes PCRProtectionProfile.AddProfileOR in favour of
PCRProtectionProfile.AddBranchPoint, which inserts a branch point in
to the current branch and returns an object that allows new branches
to be added to the branch point. There are some additional APIs that
exist solely to support chaining of function calls.

This allows the interaction between AddEFISecureBootPolicyProfile,
AddBootManagerProfile and PCRProtectionProfile to be simplified a bit.