Add a new PCRProtectionProfile API for working with branches.
This renames PCRProtectionProfile.AddProfileOR to AddBranches,
and adds a new API for working with branches in PCR profiles -
PCRProtectionProfile.AddBranchPoint, which inserts a branch point
into the current branch and returns an object that allows new branches
to be added at that point.

This helps to simplify the interaction between
AddEFISecureBootPolicyProfile, AddEFIBootManagerProfile and
PCRProtectionProfile.