Define a canonical form for PCRProtectionProfile
Computing PCR digests is currently performed in 2 steps:
- Replay the profile to compute PCR value combinations for every possible
path through the profile.
- Convert the PCR value combinations in to a PCR selection and list of
PCR digests.

This commit introduces a canonical form for PCRProtectionProfile and
converts the profile to this before the first step. This canonical form
is a profile containing a single branch point with one branch representing
each complete path through the original profile. Other than the single
branch point at the top of the profile, the canonical form contains no
other branch points.

The motivation for this is as a step towards providing a way to replay
paths through a profile for the purpose of testing. The compatibility
tests currently replay a manually created list of PCR extends, and this
is sub-optimal. Obviously, the ability to replay a profile will only be
possible for a profile that starts at the TPM's current PCR values (so,
it's not possible to replay a profile that contains AddPCRValueFromTPM, or
AddPCRValue with any value other than the reset value unless you know
what events you need to replay to get to those PCR values).