efi: Compute PK measurement digest based on the current variable value
When precomputing values for PCR7, the digest for the PK mesurement is
replayed directly from the event log even though this is a runtime
updateable variable (as long as you have an update signed with the
current PK!), so this isn't really the correct behaviour. This updates
the code to behave as follows:
- For secure boot configuration events that correspond to non-volatile
runtime variables (PK, KEK, db and dbx, and dbt and dbr if present), the
digest is computed based on the current variable value, as this value is
authoritative.
- For secure boot configuraion events that correspond to volatile
runtime variables (SecureBoot, and DeployedMode if present and disabled),
the digest is replayed directly from the event log. This is because
volatile runtime variables are mirrored by boot services code from
boot service only non-volatile variables, and the backing non-volatile
variables are not available for read or write at runtime. In this case,
the value recorded in the log is authoritative.

To test that the code picks the platform key from the correct source,
I've added a PK variable to each of the efi/testdata/efivars* directories
containing the platform key that the tests originally used from the test
event logs, and I've added an event log from my own Dell machine which
contains a different platform key, and used this log for the tests to
make sure that the computed digests didn't change.