Add internal/lukstokens package
This adds a new package containing a representation of LUKS2
tokens for use by secboot. It also contains an API (TokenView)
to provide a read-only view of tokens.

Each token has a name, which is an arbitrary string used to identify
its associated keyslot. The name is intended to be unique between
keyslots. Tokens without a name are ignored by TokenView.

A token must be associated with a single keyslot, else they are
ignored by TokenView.

The code in crypt.go that manages tokens will be expected to enforce
the invariant that the token names are unique.

For context, here's some notes about how TokenView is intended to be
used:
- TokenView.TokenByName returns the token for the keyslot with the
supplied name, and is intended to be used when performing changes to
the slot (eg, updating the PCR policy in the TPM case or deleting the
slot).
- TokenView.KeyDataTokensByPriority returns a list of key data tokens
in priority order, and is intended to be used internally by the
ActivateVolumeWithKeyData API.
- TokenView.OrphanedTokenIds returns a list of IDs associated with
tokens that have been orphaned (ie, their associated keyslot has been
deleted), and can occur if the process of deleting a keyslot and its
associated token is interrupted. This can be used to resume an
interrupted removal.