efi/preinstall: Add a HostEnvironment
This adds a HostEnvironment interface to efi/preinstall, which is a
superset of the one that's already in efi. The defaultEnvImpl in
efi/internal has been updated to provide a default implementation of it,
and the internal/efitest package has been updated to be able to mock the
new parts of it.