efi/preinstall Add initial RunChecksContext API
This API is easier to use than the other exported API (RunChecks) and
converts errors into error kinds that are suitable for representing in
JSON.

This API will eventually map certain error kinds to actions that can be
performed by this package in order to resolve some issues. The intention
here is that the installer UI eventually gives the user some information
about why FDE cannot be enabled, and it gives them the option of choosing
to fix some issues where it is possible and appropriate to do so (eg, say
the TPM device is detected but disabled, there should be an action to
enable it via the PPI).

The Action type and some pseudo actions are already defined - these are
actions that cannot be performed by this package, but are hints for
actions that can be performed by the caller, eg, shutdown, reboot and
reboot to firmware settings are some examples.

RunChecksContext will keep track of which actions are expected (based on
the previous response) and which actions are enabled - some actions are
enabled by default but may become disabled later on because of an error,
and some other actions will only be made available after performing
tests that indicate the action can be performed.